Correct min/max temperature parsing
The previous scheme computed the average over maximum expected
temperatures throughout a day and minimum expected temperatures
throughout a day.
(Essentially, under the previous scheme minimum and maximum temperatures
were daily averages that accounted for uncertainty in the forecast).
Now, minimum temperature is computed as the minimum of
forecast temperatures for a day and the maximum temperature is
computed as the maximum of forecast temperatures for a day.